# back/src/documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Beacon Management",
    "description": "Documentation"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local Dev"
    }
  ],
  "tags": [
    {
      "name": "Beacons",
      "description": "API for managing beacons"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/beacons": {
      "get": {
        "tags": [
          "Beacons"
        ],
        "summary": "Get all beacons",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Beacons"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Beacons"
        ],
        "summary": "Create a new beacon",
        "requestBody": {
          "description": "Beacon Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Beacon"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/beacon/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the beacon that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Beacons"
        ],
        "summary": "Get beacon with given ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Beacon with id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Beacon"
            }
          },
          "404": {
            "description": "Failed. Beacon not found."
          }
        }
      },
      "put": {
        "summary": "Update beacon with given ID",
        "tags": [
          "Beacons"
        ],
        "requestBody": {
          "description": "Beacon Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Beacon"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Beacon with new values of properties",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Beacon"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          },
          "404": {
            "description": "Failed. Beacon not found."
          }
        }
      },
      "delete": {
        "summary": "Delete beacon with given ID",
        "tags": [
          "Beacons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Delete Beacon with id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/id"
            }
          },
          "404": {
            "description": "Failed. Beacon not found."
          }
        }
      }
    }
  },
  "definitions": {
    "id": {
      "properties": {
        "uuid": {
          "type": "string"
        }
      }
    },
    "Beacon": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "uuid": {
          "type": "string"
        },
        "major": {
          "type": "integer"
        },
        "minor": {
          "type": "integer"
        }
      }
    },
    "Beacons": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Beacon"
      }
    }
  }
}